fix: fix build buy moving renamed component 2
