+++ src/components/index.ts
@@ -1,2 +1,2 @@
 export * from './NotFound';
-export * from './SimpleModalComponent';
+export * from './SimpleModal';
